[LCAM-2053] Refactored and renamed the PartnerService class
I've extracted much of the logic to a single method in the repository to reduce a lot of unnecessary code.
