+++ maat-court-data-api/src/main/java/gov/uk/courtdata/applicant/repository/RepOrderApplicantLinksRepository.java
@@ -1,6 +1,7 @@
 package gov.uk.courtdata.applicant.repository;
 
 import gov.uk.courtdata.applicant.entity.RepOrderApplicantLinksEntity;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,5 +11,7 @@ import java.util.List;
 public interface RepOrderApplicantLinksRepository extends JpaRepository<RepOrderApplicantLinksEntity, Integer> {
 
     List<RepOrderApplicantLinksEntity> findAllByRepId(Integer repId);
+    
+    Optional<RepOrderApplicantLinksEntity> findFirstByRepIdAndLinkDateIsNotNullAndUnlinkDateIsNull(Integer repId);
 
 }
